fix: extract shared SSE transform and fix chunk-boundary bugs
The inline TransformStream in both decomposition/extract and
formalization/lean had two bugs: (1) non-matching event blocks
re-enqueued the entire original chunk instead of just that block,
risking duplicate data, and (2) chunk boundaries were assumed to
align with SSE event boundaries, so a split mid-event produced
malformed output.

Extract the logic into transformSseStream.ts with proper buffering
and a simple onDone callback. Both routes are now ~30 lines shorter.
